backend/requirements.txt
--index-url https://pypi.org/simple
--trusted-host pypi.org
--trusted-host files.pythonhosted.org


fastapi>=0.95.2
uvicorn>=0.21.1
supabase>=1.0.0
python-dotenv>=1.0.0
httpx>=0.23.3
pydantic>=1.10.6
python-multipart>=0.0.17
pydantic-settings>=2.6.1
python-docx>=1.1.2
langchain-community>=0.3.24
openai>=1.61.1
stripe>=12.1.0
PyYAML>=5.4.1 --only-binary :all:
langchain_openai
markdown
reportlab==4.1.0
beautifulsoup4==4.12.3
python-docx
PyPDF2
langchain
openai
tiktoken
python-multipart  # For file uploads
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pypdf
python-docx
email-validator